renamed packages and removed unnecessary files
