+++ reference-project/k8s/deployment.yaml
@@ -24,7 +24,7 @@ spec:
       containers:
         - name: demo-quarkus
           # where should the image come from
-          image: 'k3d-registry:5000/demo-quarkus:latest'
+          image: 'k3d-registry.localhost:5000/demo-quarkus:latest'
           # always perform pull when creating pod
           imagePullPolicy: Always
           # env vars so we can connect to DB
