+++ reference-project/pom.xml
@@ -2,7 +2,7 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.capgemini.academy</groupId>
+  <groupId>com.devonfw.quarkus.sample</groupId>
   <artifactId>demo-quarkus</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <properties>
