+++ reference-project/src/main/java/com/devonfw/application/quarkus/sample/animalmanagement/common/api/AnimalMapper.java
@@ -1,4 +1,4 @@
-package com.capgemini.application.quarkus.sample.animalmanagement.common.api;
+package com.devonfw.application.quarkus.sample.animalmanagement.common.api;
 
 import java.util.stream.Stream;
 
@@ -7,11 +7,10 @@ import org.tkit.quarkus.jpa.daos.PageResult;
 import org.tkit.quarkus.rs.mappers.OffsetDateTimeMapper;
 import org.tkit.quarkus.rs.models.PageResultDTO;
 
-import com.capgemini.application.quarkus.sample.animalmanagement.common.api.Animal;
-import com.capgemini.application.quarkus.sample.animalmanagement.common.api.AnimalSearchCriteria;
-import com.capgemini.application.quarkus.sample.animalmanagement.logic.api.to.AnimalDTO;
-import com.capgemini.application.quarkus.sample.animalmanagement.logic.api.to.AnimalSearchCriteriaDTO;
-import com.capgemini.application.quarkus.sample.animalmanagement.logic.api.to.NewAnimalDTO;
+import com.devonfw.application.quarkus.sample.animalmanagement.dataaccess.api.Animal;
+import com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to.AnimalDTO;
+import com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to.AnimalSearchCriteriaDTO;
+import com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to.NewAnimalDTO;
 
 //mapstruct will generate an impl class from this interface at compile time
 @Mapper(uses = OffsetDateTimeMapper.class)
