+++ reference-project/src/main/java/com/devonfw/application/quarkus/sample/animalmanagement/logic/api/to/AnimalDTO.java
@@ -1,4 +1,4 @@
-package com.capgemini.application.quarkus.sample.animalmanagement.logic.api.to;
+package com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to;
 
 import lombok.Getter;
 import lombok.Setter;
