+++ reference-project/src/main/java/com/devonfw/application/quarkus/sample/animalmanagement/logic/api/to/NewAnimalDTO.java
@@ -1,4 +1,4 @@
-package com.capgemini.application.quarkus.sample.animalmanagement.logic.api.to;
+package com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to;
 
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
