+++ reference-project/src/main/resources/application.properties
@@ -9,7 +9,7 @@ quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.log.sql=true
 
 # which packages should be automatically logged by tkit cdi logger
-quarkus.tkit.log.packages=org.tkit,com.capgemini
+quarkus.tkit.log.packages=org.tkit,com.devonfw
 
 #you can supress logs from classes by annotations or regex
 quarkus.tkit.log.ignore.pattern=.*MapperImpl
@@ -44,3 +44,8 @@ quarkus.micrometer.binder.http-server.enabled=true
 
 # TEST
 %test.quarkus.tkit.log.console.json=false
+
+quarkus.datasource.db-kind=postgresql
+quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/demo
+quarkus.datasource.username=demo
+quarkus.datasource.password=demo
